cleanup 2 command return types (#13)
Almost all commands (except those that return values to the
configuration file) return nothing to the cli layer. All output
and errors are logged to the context. This change is to fix the
only two commands, "user add" and "health" that were not consistent.

testing done: unit tests and manual

  $ ./priam add user woz

  $ ./priam health